Refactored identical code for getting bookingRequest and form into sharedUtils.ts
