+++ src/routes/(app)/booking/admin/[id]/+page.server.ts
@@ -1,12 +1,11 @@
-import * as m from "$paraglide/messages";
-import { error } from "@sveltejs/kit";
-import dayjs from "dayjs";
-import { superValidate } from "sveltekit-superforms/server";
-import { zod } from "sveltekit-superforms/adapters";
-import { bookingSchema } from "../../schema";
 import { authorize } from "$lib/utils/authorization";
 import apiNames from "$lib/utils/apiNames";
-import { actions, getAllBookingRequestsWeekly } from "../sharedUtils";
+import {
+  actions,
+  getAllBookingRequestsWeekly,
+  getBookingRequestOrThrow,
+  getSuperValidatedForm,
+} from "../sharedUtils";
 
 export const load = async ({ locals, params }) => {
   const { prisma, user } = locals;
@@ -19,30 +18,8 @@ export const load = async ({ locals, params }) => {
     },
   });
 
-  const bookingRequest = await prisma.bookingRequest.findUnique({
-    where: {
-      id: params.id,
-    },
-    include: {
-      bookables: true,
-    },
-  });
-
-  if (!bookingRequest) {
-    throw error(404, m.booking_errors_notFound());
-  }
-
-  const initialData = {
-    name: bookingRequest.event ?? undefined,
-    start: bookingRequest.start
-      ? dayjs(bookingRequest.start).format("YYYY-MM-DDTHH:mm")
-      : undefined,
-    end: bookingRequest.end
-      ? dayjs(bookingRequest.end).format("YYYY-MM-DDTHH:mm")
-      : undefined,
-    bookables: bookingRequest.bookables?.map((bookable) => bookable.id),
-  };
-  const form = await superValidate(initialData, zod(bookingSchema));
+  const bookingRequest = await getBookingRequestOrThrow(prisma, params.id);
+  const form = await getSuperValidatedForm(bookingRequest);
 
   return {
     bookables,
